power: max17042 : Configure SOC & VLT thresholds for POR
The interrupt thresholds for SOC are programmed such that we get an
interrupt for every 'off' percentage change in the soc. When 0% is
reported by Fuel gauge at power-up, the lower threshold becomes -1
which is 0XFF. As a result of this, the fuel gauge interrupt line is
in pending state and blocks handling of other interrupts.

The interrupt thresholds for Voltage are programmed such that we get an
interrupt if the voltage drops below cut-off threshold. When powered up
at voltages below the cut-off threshold, fuel gauge interrupt line is in
pending state and blocks handling of other interrupts.

Change-Id: I69522d205fb16b7698853f6f24768137d4708d07
Signed-off-by: Vidyadhar Gurram <rnh786@motorola.com>
Reviewed-by: Kenneth Kessler <kennykessler@motorola.com>
Reviewed-by: Jiang-Li Yuan <jlyuan@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>