power: mmi-wls: Set Vref
Need to Set Vref to 450 mV so that the
staility of some environments is increased.
Must be set after Pad placement since I2C
block of the part does not function without
power from the Pad.  Also it must be written
each time since the register reverts to default
after Pad removal.

Change-Id: I9e1f649bc03ff03dcab19f41af0fdf50656d8387
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>